decimal needed

+++ Reports/models.py
@@ -144,8 +144,8 @@ class Report(models.Model):
     destroyed100_0 = models.IntegerField(default=2)
     destroyed100_1 = models.IntegerField(default=2)
 
-    perDam= models.IntegerField(default=-1)
-    estimated_damage= models.IntegerField(default=-1)
+    perDam= models.DecimalField(default=-1)
+    estimated_damage= models.DecimalField(default=-1)
     fema_disaster_number = models.IntegerField(default=-1)
     # normal_water = models.CharField(max_length=50)
